src/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
psycopg2-binary==2.9.10
python-dotenv==1.1.0
azure-identity==1.18.0
azure-keyvault-secrets==4.8.0